Remove only_if
Because they don't indicate composer presence correctly.
Composer is installed in /usr/local/bin by default, which is not on the path for only_if. If an execute is added before the require or install, the path is set correctly. This is most likely a bug in chef somehow, the env is different for only_if then for the execute statement.